pyproject.toml
[project]
name = "agr-blastdb-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Adam Wright, Paulo Nuin and the Alliance of Genome Resources"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "rich>=13.5.2",
    "click>=8.1.6",
    "requests>=2.31.0",
    "tqdm>=4.66.3",
    "wget>=3.2",
    "pyyaml>=6.0.1",
    "boto3>=1.28.46",
    "python-dotenv>=1.0.0",
    "slack-sdk>=3.23.0",
    "numpy>=1.26.4",
    "selenium>=4.20.0",
    "beautifulsoup4>=4.12.3",
    "webdriver-manager>=4.0.2",
]

[project.optional-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=24.3.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.12.12",
    "types-boto3>=1.0.2",
    "ruff>=0.12.8",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "locust>=2.15.1",
    "responses>=0.23.3",
    "pillow>=10.0.0",
    "psutil>=5.9.0",
    "cloudscraper>=1.2.71",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
